test(cluster): rebase the stamping tests onto conversation transcripts
CI caught these: I had run only the a2a and managed_agent crates, and
four cluster-profile integration tests write to a mailbox path to prove
`from` is unforgeable -- across trust domains, over cross-org mTLS, in
both federation directions, and for a guest confined to its own log.
Their subject is the stamp, not the path, so they move to a transcript
rather than being deleted.

Each builds the path through the a2a address SSOT instead of spelling a
leaf, so a layout change that would un-stamp those writes breaks the
test rather than leaving it green and empty. The foreign-CA pair needed
no other change: a guest may still open its own transcript without
creating the directories above it, which is the cross-org first-contact
behaviour they exist to pin.

Five policy unit tests were a quieter version of the same problem. They
assert `None` for a named reason -- the `from` already matches, no
agent_id, non-JSON -- but their path had stopped being a mailbox, so
`None` was guaranteed and the reason went untested. They use a
transcript now.

Also drops two stale `chat-with-me` strings used as arbitrary paths in
kernel unit tests, and a redundant `DT_STREAM as u8` in proc_entry that
was pre-existing but sits in a file this branch already rewrites.

+++ rust/managed_agent/src/proc_entry.rs
@@ -191,7 +191,7 @@ mod tests {
     fn stream_exists(kernel: &Kernel, path: &str) -> bool {
         kernel
             .sys_stat(path, ROOT_ZONE_ID)
-            .is_some_and(|e| e.entry_type == DT_STREAM as u8)
+            .is_some_and(|e| e.entry_type == DT_STREAM)
     }
 
     fn entry_present(kernel: &Kernel, path: &str) -> bool {
